+++ example/lib/main.dart
@@ -2,29 +2,68 @@ import 'package:easy_search_bar/easy_search_bar.dart';
 import 'package:flutter/material.dart';
 
 void main() {
-  runApp(MyHomePage());
+  runApp(const MyHomePage());
 }
 
 class MyHomePage extends StatefulWidget {
+  const MyHomePage({Key? key}) : super(key: key);
+
   @override
   State<MyHomePage> createState() => _MyHomePageState();
 }
 
 class _MyHomePageState extends State<MyHomePage> {
   String searchValue = '';
+  final List<String> _suggestions = ['Afeganistan', 'Albania', 'Algeria', 'Australia', 'Brazil', 'German', 'Madagascar', 'Mozambique', 'Portugal', 'Zambia'];
+
+  Future<List<String>> _fetchSuggestions(String searchValue) async {
+    await Future.delayed(const Duration(milliseconds: 750));
+
+    return _suggestions.where((element) {
+      return element.toLowerCase().contains(searchValue.toLowerCase());
+    }).toList();
+  }
 
   @override
   Widget build(BuildContext context) {
     return  MaterialApp(
       title: 'Example',
       theme: ThemeData(
-        primarySwatch: Colors.blue
+        primarySwatch: Colors.orange
       ),
       home: Scaffold(
-        appBar: AppBar(
-          title: EasySearchBar(
-            title: 'Example',
-            onSearch: (value) => setState(() => searchValue = value)
+        appBar: EasySearchBar(
+          title: const Text('Example'),
+          onSearch: (value) => setState(() => searchValue = value),
+          actions: [
+            IconButton(
+              icon: const Icon(
+                Icons.person,
+              ),
+              onPressed: () {}
+            )
+          ],
+          asyncSuggestions: (value) async => await _fetchSuggestions(value)
+        ),
+        drawer: Drawer(
+          child: ListView(
+            padding: EdgeInsets.zero,
+            children: [
+              const DrawerHeader(
+                decoration: BoxDecoration(
+                  color: Colors.blue,
+                ),
+                child: Text('Drawer Header'),
+              ),
+              ListTile(
+                title: const Text('Item 1'),
+                onTap: () => Navigator.pop(context)
+              ),
+              ListTile(
+                title: const Text('Item 2'),
+                onTap: () => Navigator.pop(context)
+              )
+            ]
           )
         ),
         body: Center(