+++ CHANGELOG.md
@@ -1,7 +1,19 @@
+## 2.0.0 - 16-12-2021
+
+### Added
+* Added suggestions functionality
+* Param to make appBar float
+
+### Changes
+* Updating widget so can be used as AppBar instead of being used inside AppBar title
+
 ## 1.0.1 - 02-12-2021
+
+### Changes
 
 * Updating pubspec.yaml file
 
 ## 1.0.0 - 02-12-2021
 
+### Added
 * Initial Open Source release